Add unit test for reserialization issue
When a JSchema which has remote references is serialized, the remote
references are inlined. If there are multiple refs to the same schema,
all refs after the first will point to the newly inlined reference.
Combining this with an array schema specification, schemas of this shape
will fail to deserialize.